Message redesigned to single page and design improvements
