+++ package.json
@@ -12,10 +12,13 @@
 	},
 	"dependencies": {
 		"@react-keycloak/web": "^3.4.0",
+		"axios": "^1.1.3",
 		"keycloak-js": "^20.0.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
-		"react-router-dom": "^6.4.2"
+		"react-icons": "^4.6.0",
+		"react-router-dom": "^6.4.2",
+		"react-toastify": "^9.1.1"
 	},
 	"devDependencies": {
 		"autoprefixer": "^9",
